Modelo ok, usuario post comentario

+++ src/main/java/io/redbee/blog/services/UsuarioService.java
@@ -1,20 +1,19 @@
 package io.redbee.blog.services;
-/*
+
 import io.redbee.blog.models.Usuario;
 import io.redbee.blog.repository.UsuarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-import java.util.Optional;
+import java.util.*;
 
 @Service
 public class UsuarioService {
     @Autowired
     UsuarioRepository usuarioRepository;
 
-    public ArrayList<Usuario> getUsuarios(){
-        return (ArrayList<Usuario>) usuarioRepository.findAll();
+    public List<Iterable<Usuario>> getUsuarios(){
+        return Arrays.asList(usuarioRepository.findAll());
     }
 
     public Usuario guardarUsuario(Usuario usuario){
@@ -22,19 +21,29 @@ public class UsuarioService {
     }
 
     public boolean eliminarUsuario(Long id) {
-        try {
+        Boolean encontrado = usuarioRepository.findById(id).isPresent();
+        if (encontrado) {
             usuarioRepository.deleteById(id);
-            return true;
-        }catch (Exception err){
-            return false;
         }
+        return encontrado;
 
     }
 
     public Optional<Usuario> getUsuarioPorId(Long id) {
         return usuarioRepository.findById(id);
     }
+
+    public boolean modificarUsuario(Long id, Usuario usuario) {
+        Optional<Usuario> usuarioActual = usuarioRepository.findById(id);
+        boolean actualizado = false;
+        if (usuarioActual.isPresent()){
+            Usuario usuarioAGuardar = usuarioActual.get();
+            usuarioAGuardar.setNombre(usuario.getNombre());
+            usuarioAGuardar.setCorreo(usuario.getCorreo());
+            usuarioAGuardar.setPosteos(usuario.getPosteos());
+            usuarioRepository.save(usuarioAGuardar);
+            actualizado = true;
+        }
+        return actualizado;
+    }
 }
-
-
- */
